refactor: change a few classNames
