+++ src/components/MovieAction/MovieAction.scss
@@ -1,8 +1,4 @@
-.movieAction {
-  background-color: red;
-}
-
-.info {
+.container {
   display: flex;
   flex-direction: column;
   justify-content: flex-end;
