+++ src/components/MovieDescription/MovieDescription.jsx
@@ -4,7 +4,7 @@ import PropTypes from 'prop-types';
 import styles from './MovieDescription.scss';
 
 const MovieDescription = ({ description, descriptionClass }) => (
-  <p className={`${styles.movieDescription} ${styles[descriptionClass]}`}>{ description }</p>
+  <p className={`${styles.description} ${styles[descriptionClass]}`}>{ description }</p>
 );
 
 MovieDescription.propTypes = {
