+++ src/components/MovieDescription/MovieDescription.scss
@@ -1,4 +1,4 @@
-.movieDescription {
+.description {
   width: 235px;
   padding: 14px;
   font-size: 16px;
