+++ src/components/MovieDetails/MovieDetails.scss
@@ -4,7 +4,6 @@
   height: 45vw;
   max-height: 1100px;
   padding: 70px 5% 50px 5%;
-  background-color: rgb(0, 0, 0);
   background-image: url('assets/images/junglebook.png');
   background-size: cover;
