+++ src/components/MovieDetailsRating/MovieDetailsRating.scss
@@ -1,4 +1,4 @@
-.movieDetailsRating {
+.container {
   padding-top: 10px;
 }
