+++ src/components/SearchPanel/SearchPanel.scss
@@ -9,7 +9,7 @@
   display: flex;
   align-items: flex-end;
 
-  .search__input {
+  .input {
     width: 250px;
     height: 35px;
     padding: 0 20px;
@@ -30,7 +30,7 @@
     }
   }
 
-  .search__submit {
+  .submit {
     display: inline-block;
     width: 35px;
     height: 35px;
